# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_import_upload_api_skills_import_upload_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_import_upload_api_skills_import_upload_post",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "错误响应的统一形状——所有域错误都长这样（`{\"detail\": \"一句话\"}`）。\n\n各端点在 `responses=` 里引用它，把错误路径写进 OpenAPI 契约：\n前端生成类型时才能覆盖错误体，而不是只看到成功路径。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SkillFileContent": {
        "description": "GET /api/skills/{name}/files/{path} 的响应体——UTF-8 文本内容。",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "content"
        ],
        "title": "SkillFileContent",
        "type": "object"
      },
      "SkillFileInfo": {
        "description": "技能包内单个文件的条目（预览清单用）。",
        "properties": {
          "path": {
            "description": "包内相对路径（POSIX 风格，如 SKILL.md、references/h3.md）",
            "title": "Path",
            "type": "string"
          },
          "previewable": {
            "description": "是否可通过文件内容端点预览",
            "title": "Previewable",
            "type": "boolean"
          },
          "role": {
            "description": "角色：skill=注入源 / reference=参考资料（两者可预览）；asset、script、other=不参与注入且不可预览",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "path",
          "role",
          "previewable"
        ],
        "title": "SkillFileInfo",
        "type": "object"
      },
      "SkillFileSaveRequest": {
        "additionalProperties": false,
        "description": "保存技能文本文件，并核对编辑时的原始内容。",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "original_content": {
            "title": "Original Content",
            "type": "string"
          }
        },
        "required": [
          "content",
          "original_content"
        ],
        "title": "SkillFileSaveRequest",
        "type": "object"
      },
      "SkillFilesResponse": {
        "description": "GET /api/skills/{id}/files 的响应体。",
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/SkillFileInfo"
            },
            "title": "Files",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "files"
        ],
        "title": "SkillFilesResponse",
        "type": "object"
      },
      "SkillImportRequest": {
        "description": "POST /api/skills/import 的请求体。",
        "properties": {
          "path": {
            "description": "skill 目录的本地路径（服务端可访问）",
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "SkillImportRequest",
        "type": "object"
      },
      "SkillImportResponse": {
        "description": "导入结果。",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "description": "新 skill 的稳定 ID（后续寻址用）",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "total_bytes": {
            "title": "Total Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "enabled",
          "total_bytes"
        ],
        "title": "SkillImportResponse",
        "type": "object"
      },
...
      "SkillRenameRequest": {
        "description": "POST /api/skills/{id}/rename 的请求体。",
        "properties": {
          "new_name": {
            "description": "新显示名（可改、允许重名）",
            "title": "New Name",
            "type": "string"
          }
        },
        "required": [
          "new_name"
        ],
        "title": "SkillRenameRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "AI 打标工具（发图 + 指令产出 caption，支持迭代改写）",
    "title": "Dataset Factory",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/skills/import": {
      "post": {
        "description": "从本机路径导入 skill（默认启用）：目录整包复制；单个文件按 SKILL.md 单文件导入。\n\n目录源走 agentskills.io 标准整包复制；指向一个 ``.md`` 文件时视为「无文件夹结构的\n单文件 skill」——文件整体按 SKILL.md 交付，名称 / 描述取自它的 frontmatter。\n同显示名的包可并存（身份是 ID，导入不再拒绝重名）。",
        "operationId": "import_one_api_skills_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "路径不存在 / 格式不合法"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import One",
        "tags": [
          "Skill 库"
        ]
      }
    },
    "/api/skills/import-upload": {
      "post": {
        "description": "从浏览器上传的文件集导入 skill 包（文件夹选择器 / 拖拽；传内容不传路径）。",
        "operationId": "import_upload_api_skills_import_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_upload_api_skills_import_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "上传内容不合法（缺 SKILL.md / 文件名穿越 / 格式非法）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Upload",
        "tags": [
          "Skill 库"
        ]
      }
    },
    "/api/skills/{sid}": {
      "delete": {
        "description": "删除 skill（整目录移除）。",
        "operationId": "remove_api_skills__sid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "skill 不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove",
        "tags": [
          "Skill 库"
        ]
      }
    },
    "/api/skills/{sid}/disable": {
      "post": {
        "description": "停用 skill（保留在库中，打标不注入）。",
        "operationId": "disable_api_skills__sid__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "skill 不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable",
        "tags": [
          "Skill 库"
        ]
      }
    },
    "/api/skills/{sid}/enable": {
      "post": {
        "description": "启用 skill。",
        "operationId": "enable_api_skills__sid__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "skill 不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable",
        "tags": [
          "Skill 库"
        ]
      }
    },
    "/api/skills/{sid}/files": {
      "get": {
        "description": "列出技能包内文件（角色标注：SKILL.md 与 references/ 可预览，assets / scripts 灰显占位）。",
        "operationId": "list_package_files_api_skills__sid__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillFilesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "skill 不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Package Files",
        "tags": [
          "Skill 库"
        ]
      }
    },
    "/api/skills/{sid}/files/{path}": {
      "get": {
        "description": "读技能包内一个可预览文件的文本内容（UTF-8；仅 SKILL.md 与 references/ 开放）。",
        "operationId": "read_package_file_api_skills__sid__files__path__get",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillFileContent"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "路径不合法 / 文件不参与预览 / 内容不是 UTF-8 文本"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "skill 或包内文件不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Package File",
        "tags": [
          "Skill 库"
        ]
      },
      "put": {
        "description": "写回现有技能文本文件，SKILL.md 同时校验其 frontmatter。",
        "operationId": "save_package_file_api_skills__sid__files__path__put",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillFileSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillFileContent"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "文件或内容不合法"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "文件不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "文件已被其他写者修改"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Package File",
        "tags": [
          "Skill 库"
        ]
      }
    },
    "/api/skills/{sid}/rename": {
      "post": {
        "description": "改显示名（只写 SKILL.md frontmatter 的 name 字段；引用存 ID、不受影响）。",
        "operationId": "rename_api_skills__sid__rename_post",
        "parameters": [
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "title": "Sid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillRenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "skill 不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename",
        "tags": [
          "Skill 库"
        ]
      }
    },
...
  }
}